ref: reorganize huge files tournament-management.ts and openapi meta
