+++ src/server/mutations/tournament-players.ts
@@ -1,0 +1,529 @@
+'use server';
+
+import { validateRequest } from '@/lib/auth/lucia';
+import { newid } from '@/lib/utils';
+import { db } from '@/server/db';
+import { players } from '@/server/db/schema/players';
+import { games, players_to_tournaments } from '@/server/db/schema/tournaments';
+import { getStatusInTournament } from '@/server/queries/get-status-in-tournament';
+import { getTournamentById } from '@/server/queries/tournament-helpers';
+import {
+  AddDoublesTeamModel,
+  EditDoublesTeamModel,
+  PlayerToTournamentInsertModel,
+} from '@/server/zod/tournaments';
+import {
+  PlayerFormModel,
+  PlayerInsertModel,
+  PlayerTournamentModel,
+} from '@/server/zod/players';
+import { and, eq, inArray, ne, or, sql } from 'drizzle-orm';
+import { normalizeSwissRoundsNumber } from './tournament-lifecycle';
+
+export async function removePlayer({
+  tournamentId,
+  playerId,
+  userId,
+}: {
+  tournamentId: string;
+  playerId: string;
+  userId: string;
+}) {
+  const { user } = await validateRequest();
+  if (!user) throw new Error('UNAUTHORIZED_REQUEST');
+  if (user.id !== userId) throw new Error('USER_NOT_MATCHING');
+  const tournament = await getTournamentById(tournamentId);
+  if (!tournament) throw new Error('TOURNAMENT NOT FOUND');
+  if (tournament.startedAt) throw new Error('TOURNAMENT_ALREADY_STARTED');
+
+  if (tournament.type === 'doubles') {
+    const participant = await db
+      .select({ teamNickname: players_to_tournaments.teamNickname })
+      .from(players_to_tournaments)
+      .where(
+        and(
+          eq(players_to_tournaments.tournamentId, tournamentId),
+          eq(players_to_tournaments.playerId, playerId),
+        ),
+      )
+      .then((rows) => rows.at(0));
+    if (!participant) throw new Error('TOURNAMENT_PLAYER_NOT_FOUND');
+
+    await db.transaction(async (tx) => {
+      await tx
+        .delete(games)
+        .where(
+          and(
+            eq(games.tournamentId, tournamentId),
+            or(eq(games.whiteId, playerId), eq(games.blackId, playerId)),
+          ),
+        );
+
+      if (participant.teamNickname) {
+        await tx
+          .delete(players_to_tournaments)
+          .where(
+            and(
+              eq(players_to_tournaments.tournamentId, tournamentId),
+              eq(players_to_tournaments.teamNickname, participant.teamNickname),
+            ),
+          );
+      } else {
+        await tx
+          .delete(players_to_tournaments)
+          .where(
+            and(
+              eq(players_to_tournaments.playerId, playerId),
+              eq(players_to_tournaments.tournamentId, tournamentId),
+            ),
+          );
+      }
+    });
+    await normalizeSwissRoundsNumber(tournamentId);
+    return;
+  }
+
+  await db
+    .delete(players_to_tournaments)
+    .where(
+      and(
+        eq(players_to_tournaments.playerId, playerId),
+        eq(players_to_tournaments.tournamentId, tournamentId),
+      ),
+    );
+  await normalizeSwissRoundsNumber(tournamentId);
+}
+
+export async function addNewPlayer({
+  tournamentId,
+  player,
+  addedAt,
+}: {
+  tournamentId: string;
+  player: PlayerFormModel & { id?: string };
+  addedAt?: Date;
+}) {
+  const now = addedAt ?? new Date();
+  const tournament = await getTournamentById(tournamentId);
+  if (!tournament) throw new Error('TOURNAMENT NOT FOUND');
+  if (tournament.startedAt) throw new Error('TOURNAMENT_ALREADY_STARTED');
+  if (tournament.type === 'doubles') {
+    throw new Error('DOUBLES_USE_PAIRS');
+  }
+
+  const playerId = player.id ?? newid();
+  await db
+    .insert(players)
+    .values({ ...player, lastSeenAt: new Date(), id: playerId });
+  const playerToTournament: PlayerToTournamentInsertModel = {
+    playerId,
+    tournamentId,
+    id: `${playerId}=${tournamentId}`,
+    wins: 0,
+    losses: 0,
+    draws: 0,
+    colorIndex: 0,
+    place: null,
+    isOut: null,
+    pairingNumber: null,
+    addedAt: now,
+    newRating: null,
+    newRatingDeviation: null,
+    newVolatility: null,
+  };
+  await db.insert(players_to_tournaments).values(playerToTournament);
+  await normalizeSwissRoundsNumber(tournamentId);
+}
+
+export async function addExistingPlayer({
+  tournamentId,
+  player,
+  userId,
+  addedAt,
+}: {
+  tournamentId: string;
+  player: PlayerInsertModel;
+  userId: string;
+  addedAt?: Date;
+}) {
+  const now = addedAt ?? new Date();
+  const { user } = await validateRequest();
+  if (!user) throw new Error('UNAUTHORIZED_REQUEST');
+  if (user.id !== userId) throw new Error('USER_NOT_MATCHING');
+  const tournament = await getTournamentById(tournamentId);
+  if (!tournament) throw new Error('TOURNAMENT NOT FOUND');
+  if (tournament.startedAt) throw new Error('TOURNAMENT_ALREADY_STARTED');
+  if (tournament.type === 'doubles') {
+    throw new Error('DOUBLES_USE_PAIRS');
+  }
+  const { status } = await getStatusInTournament(user.id, tournamentId);
+  if (status === 'viewer') throw new Error('NOT_ADMIN');
+
+  const playerToTournament: PlayerToTournamentInsertModel = {
+    playerId: player.id,
+    tournamentId: tournamentId,
+    id: `${player.id}=${tournamentId}`,
+    wins: 0,
+    losses: 0,
+    draws: 0,
+    colorIndex: 0,
+    place: null,
+    isOut: null,
+    pairingNumber: null,
+    addedAt: now,
+    newRating: null,
+    newRatingDeviation: null,
+    newVolatility: null,
+  };
+  await db.insert(players_to_tournaments).values(playerToTournament);
+  await normalizeSwissRoundsNumber(tournamentId);
+}
+
+export async function addDoublesTeam({
+  tournamentId,
+  nickname,
+  firstPlayerId,
+  secondPlayerId,
+  addedAt,
+}: AddDoublesTeamModel & {
+  tournamentId: string;
+  addedAt?: Date;
+}): Promise<PlayerTournamentModel> {
+  const now = addedAt ?? new Date();
+  const { user } = await validateRequest();
+  if (!user) throw new Error('UNAUTHORIZED_REQUEST');
+
+  if (firstPlayerId === secondPlayerId) {
+    throw new Error('INVALID_DOUBLES_PAIR');
+  }
+
+  const tournament = await getTournamentById(tournamentId);
+  if (!tournament) throw new Error('TOURNAMENT NOT FOUND');
+  if (tournament.startedAt) throw new Error('TOURNAMENT_ALREADY_STARTED');
+  if (tournament.type !== 'doubles') throw new Error('NOT_DOUBLES_TOURNAMENT');
+
+  const selectedPlayers = await db
+    .select({
+      id: players.id,
+      nickname: players.nickname,
+      rating: players.rating,
+    })
+    .from(players)
+    .where(
+      and(
+        eq(players.clubId, tournament.clubId),
+        or(eq(players.id, firstPlayerId), eq(players.id, secondPlayerId)),
+      ),
+    );
+
+  if (selectedPlayers.length !== 2) {
+    throw new Error('PAIR_PLAYERS_NOT_FOUND');
+  }
+
+  const existingPair = await db
+    .select({ id: players_to_tournaments.id })
+    .from(players_to_tournaments)
+    .where(
+      and(
+        eq(players_to_tournaments.tournamentId, tournamentId),
+        or(
+          eq(players_to_tournaments.playerId, firstPlayerId),
+          eq(players_to_tournaments.playerId, secondPlayerId),
+        ),
+      ),
+    )
+    .limit(1);
+
+  if (existingPair.length > 0) {
+    throw new Error('PLAYER_ALREADY_IN_PAIR');
+  }
+
+  const existingNickname = await db
+    .select({ id: players_to_tournaments.id })
+    .from(players_to_tournaments)
+    .where(
+      and(
+        eq(players_to_tournaments.tournamentId, tournamentId),
+        eq(
+          sql<string>`lower(${players_to_tournaments.teamNickname})`,
+          nickname.toLowerCase(),
+        ),
+      ),
+    )
+    .limit(1);
+
+  if (existingNickname.length > 0) {
+    throw new Error('PAIR_NICKNAME_TAKEN');
+  }
+
+  const selectedPlayersById = new Map(
+    selectedPlayers.map((each) => [each.id, each]),
+  );
+  const orderedPlayers = [firstPlayerId, secondPlayerId].map((id) => {
+    const player = selectedPlayersById.get(id);
+    if (!player) throw new Error('PAIR_PLAYERS_NOT_FOUND');
+    return player;
+  });
+  const leaderPlayerId = firstPlayerId;
+
+  const teamRating = Math.round(
+    orderedPlayers.reduce((acc, player) => acc + player.rating, 0) /
+      orderedPlayers.length,
+  );
+
+  const teamMembers: PlayerToTournamentInsertModel[] = [
+    {
+      playerId: firstPlayerId,
+      tournamentId,
+      id: `${firstPlayerId}=${tournamentId}`,
+      wins: 0,
+      losses: 0,
+      draws: 0,
+      colorIndex: 0,
+      place: null,
+      isOut: null,
+      pairingNumber: null,
+      teamNickname: nickname,
+      numberInTeam: 1,
+      addedAt: now,
+      newRating: null,
+      newRatingDeviation: null,
+      newVolatility: null,
+    },
+    {
+      playerId: secondPlayerId,
+      tournamentId,
+      id: `${secondPlayerId}=${tournamentId}`,
+      wins: 0,
+      losses: 0,
+      draws: 0,
+      colorIndex: 0,
+      place: null,
+      isOut: null,
+      pairingNumber: null,
+      teamNickname: nickname,
+      numberInTeam: 2,
+      addedAt: now,
+      newRating: null,
+      newRatingDeviation: null,
+      newVolatility: null,
+    },
+  ];
+
+  await db.insert(players_to_tournaments).values(teamMembers);
+
+  await normalizeSwissRoundsNumber(tournamentId);
+
+  return {
+    id: leaderPlayerId,
+    nickname,
+    realname: null,
+    rating: teamRating,
+    wins: 0,
+    draws: 0,
+    losses: 0,
+    colorIndex: 0,
+    isOut: null,
+    place: null,
+    pairingNumber: null,
+    addedAt: now,
+    teamNickname: nickname,
+    username: null,
+    pairPlayers: orderedPlayers.map((player) => ({
+      id: player.id,
+      nickname: player.nickname,
+    })),
+  };
+}
+
+export async function editDoublesTeam({
+  tournamentId,
+  currentTeamPlayerId,
+  nickname,
+  firstPlayerId,
+  secondPlayerId,
+}: EditDoublesTeamModel & {
+  tournamentId: string;
+}): Promise<void> {
+  const { user } = await validateRequest();
+  if (!user) throw new Error('UNAUTHORIZED_REQUEST');
+
+  if (firstPlayerId === secondPlayerId) {
+    throw new Error('INVALID_DOUBLES_PAIR');
+  }
+
+  const tournament = await getTournamentById(tournamentId);
+  if (!tournament) throw new Error('TOURNAMENT NOT FOUND');
+  if (tournament.startedAt) throw new Error('TOURNAMENT_ALREADY_STARTED');
+  if (tournament.type !== 'doubles') throw new Error('NOT_DOUBLES_TOURNAMENT');
+
+  const participant = await db
+    .select({
+      teamNickname: players_to_tournaments.teamNickname,
+      addedAt: players_to_tournaments.addedAt,
+    })
+    .from(players_to_tournaments)
+    .where(
+      and(
+        eq(players_to_tournaments.tournamentId, tournamentId),
+        eq(players_to_tournaments.playerId, currentTeamPlayerId),
+      ),
+    )
+    .then((rows) => rows.at(0));
+
+  if (!participant?.teamNickname) {
+    throw new Error('TOURNAMENT_PLAYER_NOT_FOUND');
+  }
+  const currentTeamNickname = participant.teamNickname;
+  const preservedAddedAt = participant.addedAt ?? new Date();
+
+  const selectedPlayers = await db
+    .select({
+      id: players.id,
+      nickname: players.nickname,
+      rating: players.rating,
+    })
+    .from(players)
+    .where(
+      and(
+        eq(players.clubId, tournament.clubId),
+        or(eq(players.id, firstPlayerId), eq(players.id, secondPlayerId)),
+      ),
+    );
+
+  if (selectedPlayers.length !== 2) {
+    throw new Error('PAIR_PLAYERS_NOT_FOUND');
+  }
+
+  const currentTeamMembers = await db
+    .select({ playerId: players_to_tournaments.playerId })
+    .from(players_to_tournaments)
+    .where(
+      and(
+        eq(players_to_tournaments.tournamentId, tournamentId),
+        eq(players_to_tournaments.teamNickname, currentTeamNickname),
+      ),
+    );
+
+  const currentTeamMemberIds = new Set(
+    currentTeamMembers.map((member) => member.playerId),
+  );
+
+  const occupiedPlayers = await db
+    .select({
+      playerId: players_to_tournaments.playerId,
+      teamNickname: players_to_tournaments.teamNickname,
+    })
+    .from(players_to_tournaments)
+    .where(
+      and(
+        eq(players_to_tournaments.tournamentId, tournamentId),
+        inArray(players_to_tournaments.playerId, [
+          firstPlayerId,
+          secondPlayerId,
+        ]),
+      ),
+    );
+
+  const hasOtherTeamMember = occupiedPlayers.some(
+    (row) =>
+      !currentTeamMemberIds.has(row.playerId) &&
+      row.teamNickname !== currentTeamNickname,
+  );
+  if (hasOtherTeamMember) {
+    throw new Error('PLAYER_ALREADY_IN_PAIR');
+  }
+
+  const existingNickname = await db
+    .select({ id: players_to_tournaments.id })
+    .from(players_to_tournaments)
+    .where(
+      and(
+        eq(players_to_tournaments.tournamentId, tournamentId),
+        eq(
+          sql<string>`lower(${players_to_tournaments.teamNickname})`,
+          nickname.toLowerCase(),
+        ),
+        ne(players_to_tournaments.teamNickname, currentTeamNickname),
+      ),
+    )
+    .limit(1);
+
+  if (existingNickname.length > 0) {
+    throw new Error('PAIR_NICKNAME_TAKEN');
+  }
+
+  await db.transaction(async (tx) => {
+    await tx
+      .delete(players_to_tournaments)
+      .where(
+        and(
+          eq(players_to_tournaments.tournamentId, tournamentId),
+          eq(players_to_tournaments.teamNickname, currentTeamNickname),
+        ),
+      );
+
+    const selectedPlayersById = new Map(
+      selectedPlayers.map((each) => [each.id, each]),
+    );
+    const orderedPlayers = [firstPlayerId, secondPlayerId].map((id) => {
+      const player = selectedPlayersById.get(id);
+      if (!player) throw new Error('PAIR_PLAYERS_NOT_FOUND');
+      return player;
+    });
+
+    const [firstPlayer, secondPlayer] = orderedPlayers;
+
+    await tx.insert(players_to_tournaments).values([
+      {
+        playerId: firstPlayer.id,
+        tournamentId,
+        id: `${firstPlayer.id}=${tournamentId}`,
+        wins: 0,
+        losses: 0,
+        draws: 0,
+        colorIndex: 0,
+        place: null,
+        isOut: null,
+        pairingNumber: null,
+        teamNickname: nickname,
+        numberInTeam: 1,
+        addedAt: preservedAddedAt,
+        newRating: null,
+        newRatingDeviation: null,
+        newVolatility: null,
+      },
+      {
+        playerId: secondPlayer.id,
+        tournamentId,
+        id: `${secondPlayer.id}=${tournamentId}`,
+        wins: 0,
+        losses: 0,
+        draws: 0,
+        colorIndex: 0,
+        place: null,
+        isOut: null,
+        pairingNumber: null,
+        teamNickname: nickname,
+        numberInTeam: 2,
+        addedAt: preservedAddedAt,
+        newRating: null,
+        newRatingDeviation: null,
+        newVolatility: null,
+      },
+    ]);
+  });
+
+  await normalizeSwissRoundsNumber(tournamentId);
+}
+
+export async function resetTournamentPlayers({
+  tournamentId,
+}: {
+  tournamentId: string;
+}) {
+  await db.delete(games).where(eq(games.tournamentId, tournamentId));
+  await db
+    .delete(players_to_tournaments)
+    .where(eq(players_to_tournaments.tournamentId, tournamentId));
+}
